Add better handling for checkout interruptions
If you lose your connection to vCenter while putting a host into
maintenance mode, that task will eventually complete, but then the host
won't be moved into the right cluster. Even worse, re-running the
command will just error.

This change fixes it so that re-running the checkout command will
properly detect when the chosen host is already in maintenance mode, and
skip ahead to moving the host and bringing it up again.